requirements.txt
annotated-types
anyio
asgiref
attrs
backoff
bcrypt
build
cachetools
certifi
charset-normalizer
chromadb
click
colorama
coloredlogs
crispy-bootstrap5
distlib
distro
Django
django-crispy-forms
django-environ
djangorestframework
durationpy
fastapi
filelock
flatbuffers
fsspec
google-auth
googleapis-common-protos
greenlet
grpcio
h11
httpcore
httptools
httpx
huggingface-hub
humanfriendly
idna
importlib_metadata
importlib_resources
Jinja2
jiter
jsonpatch
jsonpointer
jsonschema
jsonschema-specifications
kubernetes
langchain
langchain-core
langchain-text-splitters
langsmith
markdown-it-py
MarkupSafe
mdurl
mmh3
MouseInfo
mpmath
numpy
oauthlib
onnxruntime
openai
opencv-python
opentelemetry-api
opentelemetry-exporter-otlp-proto-common
opentelemetry-exporter-otlp-proto-grpc
opentelemetry-proto
opentelemetry-sdk
opentelemetry-semantic-conventions
orjson
overrides
packaging
pillow
pipenv
platformdirs
pluggy
posthog
protobuf
psycopg2
pyasn1
pyasn1_modules
PyAutoGUI
pybase64
pydantic
pydantic_core
PyGetWindow
Pygments
PyMsgBox
pyperclip
PyPika
pyproject_hooks
pyreadline3
PyRect
PyScreeze
pyscript
python-dateutil
python-dotenv
python-multipart
pytweening
PyYAML
referencing
requests
requests-oauthlib
requests-toolbelt
rich
rpds-py
rsa
setuptools
six
sniffio
SQLAlchemy
sqlparse
starlette
sympy
tenacity
tokenizers
toml
tqdm
typer
typing-inspection
typing_extensions
tzdata
urllib3
uvicorn
virtualenv
watchfiles
websocket-client
websockets
zipp
zstandard
dj-database-url
psycopg2-binary